Add tests for the new registration page
