+++ pages/register.py
@@ -17,16 +17,11 @@ class Register(Base):
 
     _error_locator = (By.CSS_SELECTOR, 'div.alert.alert-error')
     _full_name_field_locator = (By.ID, 'id_full_name')
-    _website_field_locator = (By.ID, 'id_website')
-    _bio_field_locator = (By.ID, 'id_bio')
-    _skills_field_locator = (By.CSS_SELECTOR, '#id_skills + ul input')
-    _language_locator = (By.ID, 'id_language_set-0-code')
-    _previous_button_locator = (By.ID, 'page1button')
     _map_search_box_locator = (By.ID, 'location_search')
     _country_locator = (By.ID, 'display_country')
     _privacy_locator = (By.ID, 'id_optin')
     _privacy_error_message_locator = (By.CSS_SELECTOR, '.error-message')
-    _create_profile_button_locator = (By.CSS_SELECTOR, '#edit-controls > button')
+    _create_profile_button_locator = (By.CSS_SELECTOR, '#form-submit-top')
 
     @property
     def error_message(self):
@@ -42,28 +37,6 @@ class Register(Base):
     def set_full_name(self, full_name):
         element = self.selenium.find_element(*self._full_name_field_locator)
         element.send_keys(full_name)
-
-    def set_website(self, website):
-        element = self.selenium.find_element(*self._website_field_locator)
-        element.send_keys(website)
-
-    def set_bio(self, biography):
-        element = self.selenium.find_element(*self._bio_field_locator)
-        element.send_keys(biography)
-
-    def select_language(self, language):
-        element = self.selenium.find_element(*self._language_locator)
-        select = Select(element)
-        select.select_by_value(language)
-
-    def add_skill(self, skill_name):
-        element = self.selenium.find_element(*self._skills_field_locator)
-        element.send_keys(skill_name)
-        # send tab to make the entry "stick"
-        element.send_keys("\t")
-
-    def click_previous_button(self):
-        self.selenium.find_element(*self._previous_button_locator).click()
 
     @property
     def privacy_error_message(self):
